% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFSD.R
\name{area.btwn.cdfs.calc}
\alias{area.btwn.cdfs.calc}
\title{Calculates area between CDFs}
\usage{
area.btwn.cdfs.calc(sd.obj)
}
\arguments{
\item{sd.obj}{StochasticDominance object.}
}
\value{
A numeric vector, including the area between the segments of CDFs.
}
\description{
It calculates the area between the CDFs of two prospects divided by the CDFs
and outcomes segments.
}
