% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Screening.R
\name{asd.screen}
\alias{asd.screen}
\title{Create the inefficient set by an ASD rule}
\usage{
asd.screen(data, test, epsilon.threshold)
}
\arguments{
\item{data}{A data frame, including variable pairs, distributions, and
results of the tests.}

\item{test}{A string that indicates the name of the test.}

\item{epsilon.threshold}{A number that indicates the threshold for considering
the result to be valid or not.}
}
\value{
A character vector as the inefficient set of variables based on the
`test` result and `epsilon.threshold`.
}
\description{
It uses the test result, corresponding epsilon, and the epsilon threshold to
create the inefficient set.
}
\details{
The inefficient set includes all variable names that are dominated by at
least one other variable. The domination defines as the test result where
the corresponding epsilon is smaller than or equal to epsilon threshold
(`epsilon`). If the epsilon is smaller than the threshold, it is assumed that
by some relaxation the domination exists because a very small ratio of the
decision-makers do not agree with this result.
}
