% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{calc.area.below.line}
\alias{calc.area.below.line}
\title{Calculates the area between x-axis and a straight line}
\usage{
calc.area.below.line(x1, x2, y1, y2)
}
\arguments{
\item{x1, x2}{Float values, indicating the first coordinates of start and end
points, respectively.}

\item{y1, y2}{Float values, indicating the second coordinates of start and end
points, respectively.}
}
\value{
A float.
}
\description{
It takes four parameters as the coordinates of two points at the beginning
and end of a straight line, and calculates the area exactly below the line
(between the line and the x-axis).
}
