% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFSD.R
\name{calc.betw.cdf.area}
\alias{calc.betw.cdf.area}
\title{Calculates area between CDFs}
\usage{
calc.betw.cdf.area(outcome, cdf1, cdf2)
}
\arguments{
\item{outcome}{A numeric vector, including all outcomes in ascending order.}

\item{cdf1, cdf2}{Numeric vectors, including marginal CDF corresponding to
each prospect.}
}
\value{
A numeric vector, including the area between the segments of CDFs.
}
\description{
It calculates the area between the CDFs of two prospects as the segments
of the combined CDFs, which is matched with the outcomes.
}
\examples{
outcome1 = c(1,4,7)
outcome2 = c(2,3,5)
prob1 = c(1/3,1/3,1/3)
prob2 = c(1/6,1/6,2/3)
obj = fsd(outcome1, outcome2, prob1, prob2)
calc.betw.cdf.area(obj$outcome, obj$cdf1, obj$cdf2)

}
\seealso{
[fsd()] for the parameters.
}
