% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{calc.intersection}
\alias{calc.intersection}
\title{Calculates the intersection point of two lines.}
\usage{
calc.intersection(x1, x2, y11, y12, y21, y22)
}
\arguments{
\item{x1, x2}{Float values.}

\item{y11, y12}{Float values, corresponding to the first line.}

\item{y21, y22}{Float values, corresponding to the second line.}
}
\value{
A list, including the coordinates of the intersection point.
}
\description{
Given the start and end coordinates of two straight lines, we calculate the
intersection point of them.
}
\details{
Having intersection point has been checked before.
}
\seealso{
[has.intersect()]
}
