% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{calc.mod.outcome}
\alias{calc.mod.outcome}
\title{Modify outcome vector.}
\usage{
calc.mod.outcome(outcome, ssd1, ssd2)
}
\arguments{
\item{outcome}{A numeric vector.}

\item{ssd1, ssd2}{Numeric vectors.}
}
\value{
A list of three elements.
}
\description{
It modify outcome vector to have all original values plus all intersection
points in ascending order. The corresponding SSD vectors will also be
returned.
}
\examples{
outcome1 = c(1,4,7)
outcome2 = c(2,3,5)
prob1 = c(1/3,1/3,1/3)
prob2 = c(1/6,1/6,2/3)
obj.fsd = fsd(outcome1, outcome2, prob1, prob2)
obj.ssd = ssd(obj.fsd$outcome, obj.fsd$cdf1, obj.fsd$cdf2)
calc.mod.outcome(obj.fsd$outcome, obj.ssd$ssd1, obj.ssd$ssd2)

}
