% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comparisons.R
\name{compare.all}
\alias{compare.all}
\title{Comparing all pairs by all rules, and finding the sets}
\usage{
compare.all(
  variable,
  probability,
  outcome,
  afsd.epsilon.threshold = 0.1,
  assd.ll.epsilon.threshold = 0.1,
  assd.ths.epsilon.threshold = 0.1,
  include.details = TRUE
)
}
\arguments{
\item{variable}{A character vector containing the name of all variables.}

\item{probability}{A numeric vector containing the probability for each
variable to achieve a particular outcome.}

\item{outcome}{A numeric vector containing the outcome values.}

\item{afsd.epsilon.threshold}{A number that shows the upper limit for epsilon
of afsd rule. If the epsilon is smaller than or equal to this value, the result
is accepted.}

\item{assd.ll.epsilon.threshold}{A number that shows the upper limit for epsilon
of assd.ll rule. If the epsilon is smaller than or equal to this value, the result
is accepted.}

\item{assd.ths.epsilon.threshold}{A number that shows the upper limit for epsilon
of assd.ths rule. If the epsilon is smaller than or equal to this value, the result
is accepted.}

\item{include.details}{A logical that affects the output data to contain
distributions.}
}
\value{
A list of 6 elements, including a data frame and the efficient and
inefficient sets of all rules.
}
\description{
It creates a data frame of unique unordered pairs of variables, collects
corresponding distributions, tests all SD and ASD rules on every pairs, and
finds the efficient and inefficient sets for all those rules.
}
\details{
If the type of each input parameter is incorrect, it will raise an error.

The length of the input parameters `variable`, `probability`, and `outcome`
must be equal, otherwise it will raise an error.

The output is a list that contains 6 elements: `data` is a data frame including
all unique unordered pairs of variables, corresponding distributions, result
of each rule, and epsilon of ASD rule. The rest 5 elements are lists as well,
corresponding to each rule that includes the efficient and inefficient sets.

The `data` is a nested data frame, and stores the distribution of each variable
in a single cell. The distribution contains two columns: probability and
outcome.
}
\examples{
result = compare.all(data_ex$gen, rep(1/29,377), data_ex$yield)

}
\seealso{
[create.paired.distributions(), sd.asd.test.all(), screen()]
}
