% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comparisons.R
\name{compare.paired.distributions}
\alias{compare.paired.distributions}
\title{Creating paired distributions and execute all SD and ASD rules on them}
\usage{
compare.paired.distributions(
  variable,
  probability,
  outcome,
  include.details = TRUE
)
}
\arguments{
\item{variable}{A character vector containing the name of all variables.}

\item{probability}{A numeric vector containing the probability for each
variable to achieve a particular outcome.}

\item{outcome}{A numeric vector containing the outcome values.}

\item{include.details}{A logical that affects the output data to contain
distributions. The default value is `TRUE`.}
}
\value{
A data frame, including all distribution pairs and the results of
the tests.
}
\description{
It creates a data frame including all unordered pairs of the variables, where
the first variable has higher or equal mean outcome. Then, all available SD
and ASD rules are tested on them.
}
\details{
If the type of each input parameter is incorrect, it will raise an error.

The length of the input parameters `variable`, `probability`, and `outcome`
must be equal, otherwise it will raise an error.
}
\examples{
result = compare.paired.distributions(data_ex$gen, rep(1/29,377), data_ex$yield)

}
\seealso{
[create.paired.distributions(), sd.asd.test.all()]
}
