% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{comparison}
\alias{comparison}
\title{Comparing two numeric vectors}
\usage{
comparison(x, y)
}
\arguments{
\item{x, y}{Numeric vectors.}
}
\value{
An integer, among 0, 1, 2.
}
\description{
This function compares two numeric vectors. The vector whose all elements
smaller or equal to all elements of the other one where at least one element
is smaller, will be the winner.
}
\details{
The function returns the index of the winner, meaning 1 or 2, corresponding
to parameters `x` and `y`, respectively. If no vector meets the condition, 0
will be returned.
}
