% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pair_Distributions.R
\name{create.dataframe}
\alias{create.dataframe}
\title{create a data frame from input parameters}
\usage{
create.dataframe(variable, probability, outcome)
}
\arguments{
\item{variable}{A character vector.}

\item{probability}{A numeric vector.}

\item{outcome}{A numeric vector.}
}
\value{
A data frame with three columns.
}
\description{
create a data frame from input parameters
}
