% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pair_Distributions.R
\name{create.paired.distributions}
\alias{create.paired.distributions}
\title{Creating paired distributions}
\usage{
create.paired.distributions(variable, probability, outcome)
}
\arguments{
\item{variable}{A character vector.}

\item{probability}{A numeric vector.}

\item{outcome}{A numeric vector.}
}
\value{
A data frame with four columns.
}
\description{
with the input parameters and four other functions, it will create a data frame
that includes all pairs of variables and their corresponding probability and
outcome values.
}
\details{
Here, we first sort variables based on their outcome in ascending order, and
then create unordered unique pairs. So, the first element (variable) of any
pair has larger mean outcome than the second element.

The output has four columns. The first two represent the two elements of each
pair. The last two are the corresponding probability and outcome for the
elements, respectively. These are in nested format, meaning each cell of the
last two columns includes a data frame with two columns: probability and
outcome.
}
\seealso{
[create.dataframe(), sort.variables(), pair.variables(), pair.distributions()]
}
