% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{createDistributions}
\alias{createDistributions}
\title{Constructor of Distributions Class}
\usage{
createDistributions(outcome1, outcome2, prob1, prob2)
}
\arguments{
\item{outcome1, outcome2}{Numeric vectors. The outcomes corresponding to each
prospect.}

\item{prob1, prob2}{Numeric vectors. The probabilities corresponding to each
prospect.}
}
\value{
Distributions object.
}
\description{
It is much easier to use this constructor to create an instance of the
Distributions class.
}
