% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{createStochasticDominance}
\alias{createStochasticDominance}
\title{Constructor of StochasticDominance Class}
\usage{
createStochasticDominance(outcome1, outcome2, prob1, prob2)
}
\arguments{
\item{outcome1, outcome2}{Numeric vectors. The outcomes corresponding to each
prospect.}

\item{prob1, prob2}{Numeric vectors. The probabilities corresponding to each
prospect.}
}
\value{
StochasticDominance object.
}
\description{
It is much easier to use this constructor to create an instance of the
StochasticDominance class. It handles calculation of the cdf and ssd values
in an efficient way.
}
\examples{
createStochasticDominance(outcome1 = c(1,4,7),
                          outcome2 = c(2,3,5),
                          prob1 = c(1/3,1/3,1/3),
                          prob2 = c(1/6,1/6,2/3))

}
\seealso{
[StochasticDominance()]
}
