% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_ex}
\alias{data_ex}
\title{The example dataset}
\format{
A data frame with 377 rows and 3 columns:
\describe{
   \item{gen}{Character vector of genotype (cultivar) names}
   \item{env}{Character vector of environment names}
   \item{yield}{Numeric vector of yield of planting each cultivar in each
   environment}
}
}
\source{
The phenotypic data from G2F initiative in 2018:
<https://doi.org/10.25739/anqq-sg86>
}
\usage{
data_ex
}
\description{
It is a real dataset in agriculture, which includes planted cultivars in
multi-environment and the resulting yield.
}
\keyword{datasets}
