% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\docType{class}
\name{Distributions-class}
\alias{Distributions-class}
\title{Distributions Class}
\description{
Represents two distributions (prospects) that are going to be compared using
Stochastic Dominance (SD).
}
\details{
It contains the input validation needed for comparing two prospects. For example,
having sorted `outcome`, each of `prob1` and `prob2` adds up to one, arguments
having the same lengths, and having corresponding probability and cumulative
arguments.
}
\section{Slots}{

\describe{
\item{\code{outcome}}{Numeric vector. The combined outcome values in ascending order.}

\item{\code{prob1}}{Numeric vector. Corresponding probability vector for prospect one.}

\item{\code{prob2}}{Numeric vector. Corresponding probability vector for prospect two.}

\item{\code{cum.prob1}}{Numeric vector. Corresponding cumulative probability vector for prospect one.}

\item{\code{cum.prob2}}{Numeric vector. Corresponding cumulative probability vector for prospect two.}
}}

