% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{expected.values}
\alias{expected.values}
\title{Calculating Expected value}
\usage{
expected.values(sd.obj)
}
\arguments{
\item{sd.obj}{StochasticDominance object.}
}
\value{
A list, including two double elements as the expected value of each
prospect.
}
\description{
It calculates the expected value of both prospects given their distributions.
}
