% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{has.intersection}
\alias{has.intersection}
\title{If two lines have intersection or no.}
\usage{
has.intersection(x1, x2, y11, y12, y21, y22)
}
\arguments{
\item{x1, x2}{Float values.}

\item{y11, y12}{Float values, corresponding to the first line.}

\item{y21, y22}{Float values, corresponding to the second line.}
}
\value{
A Boolean.
}
\description{
It determines if two lines have intersection point or not. The start and
end points of both lines have equal first coordinate value (the one
corresponds to the x-axis).
}
