% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSD.R
\name{jointdist_calc}
\alias{jointdist_calc}
\title{Calculate the joint distribution}
\usage{
jointdist_calc(outcome1, outcome2, prob1, prob2)
}
\description{
This function takes two random prospects by their outcomes and
 probabilities, and calculates the joint distributions in ascending order.
}
\details{
@details
 The parameters that end with the same number are corresponding to the same
 prospect, e.g., `outcome1` and `prob1` are the outcome and probability
 vectors of the first prospect.

 The length of `outcome1` and `prob1` should be the same. This is true for
 the other two parameters. If not, the function raise an error.

 The `prob1` and `prob2`, each must adds up to one. If not, the function
 raise an error.

 The output includes three elements: outcome: a combined version of
 `outcome1` and `outcome2` in ascending order, cdf1, cdf2: the marginal
 cumulative distribution function (CDF) of the corresponding prospect.

 @param outcome1,outcome2 Numeric vectors, including outcome values.
 @param prob1,prob2 Numeric vectors, including probability values.
 @returns A list of three numeric vectors.
 @examples
 outcome1 = c(1,4,7)
 outcome2 = c(2,3,5)
 prob1 = c(1/3,1/3,1/3)
 prob2 = c(1/6,1/6,2/3)
 joindist_calc(outcome1, outcome2, prob1, prob2)
}
