% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pair_Distributions.R
\name{pair.distributions}
\alias{pair.distributions}
\title{Paring all the distributions}
\usage{
pair.distributions(org.data, paired.vars)
}
\arguments{
\item{org.data}{A data frame that includes all variables, probabilities, and
outcomes.}

\item{paired.vars}{A data frame that includes all unique pairs of variables.}
}
\value{
A data frame that includes all unique pairs of variables with their
corresponding distributions in the nested format.
}
\description{
It uses paired variables and collects all the related information (probability
and outcome) corresponding to each variable of every pairs.
}
