% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pair_Distributions.R
\name{pair.variables}
\alias{pair.variables}
\title{Paring all the variables}
\usage{
pair.variables(variables)
}
\arguments{
\item{variables}{A character vector, including the variable names.}
}
\value{
A data frame, that includes two columns; every row represents a pair
of variables.
}
\description{
It uses a vector of variables and creates unique unordered pairs.
}
\details{
These pairs are unique unordered pairs, meaning the order of variables does
not matter. So if we have pair \{x,y\} we do not create another pair \{y,x\},
because they are technically the same. If the number of variables is n, the
number of pairs is n(n-1)/2.
}
