% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot_fsd}
\alias{plot_fsd}
\title{Drawing the CDFs}
\usage{
plot_fsd(outcome, cdf1, cdf2, names = c("1", "2"))
}
\arguments{
\item{outcome}{A numeric vector, including all outcome values in ascending
order.}

\item{cdf1, cdf2}{Numeric vectors, including the CDF values for corresponding
prospects.}

\item{names}{A character vector, including the names of prospects in order.}
}
\value{
A list, including plot elements.
}
\description{
It visualizes the CDFs of both prospects.
}
\details{
The length of all three parameters, `outcome`, `cdf1`, and `cdf2` must be
equal. Otherwise, an error will be raised.

The parameter `names` only accepts character vector, otherwise an error will
be raised.

The function shows a step plot that includes both CDFs, and return the plot
object for further modifications.
}
\examples{
outcome1 = c(1,4,7)
outcome2 = c(2,3,5)
prob1 = c(1/3,1/3,1/3)
prob2 = c(1/6,1/6,2/3)
obj = fsd(outcome1, outcome2, prob1, prob2)
plot_fsd(obj$outcome, obj$cdf1, obj$cdf2)

}
\seealso{
[fsd()] for the parameters.
}
