% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASSD_LL.R
\name{pos.neg.area.assd.ll}
\alias{pos.neg.area.assd.ll}
\title{Calculates positive and negative area between CDFs for ASSD-LL}
\usage{
pos.neg.area.assd.ll(sd.obj)
}
\arguments{
\item{sd.obj}{StochasticDominance object.}
}
\value{
A list, including three elements corresponding to the positive and
negative areas, respectively, and the area vector at the end.
}
\description{
It calculates the positive and negative areas between CDFs. The positive is
where both CDF and SSD of the first prospect is larger, and vice versa for
the negative case.
}
\seealso{
[modif.outcome.ssd.calc()] for more details.
}
