% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSSD.R
\name{rssd.calc}
\alias{rssd.calc}
\title{Calculates RSSD values}
\usage{
rssd.calc(ssd1, ssd2)
}
\arguments{
\item{ssd1, ssd2}{Numeric vectors, including the SSD values.}
}
\value{
A list of two numbers, indicating the RSSD values.
}
\description{
It takes SSD vectors and add them up separately to calculate the RSSD values.
}
\examples{
dists = createDistributions(outcome1 = c(1,4,7),
                            outcome2 = c(2,3,5),
                            prob1 = c(1/3,1/3,1/3),
                            prob2 = c(1/6,1/6,2/3))
ssd.obj = ssd.calc(dists)
rssd.calc(ssd.obj$ssd1, ssd.obj$ssd2)

}
