% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSSD.R
\name{rssd.test}
\alias{rssd.test}
\title{Compares two random prospects by RSSD}
\usage{
rssd.test(ssd1, ssd2)
}
\arguments{
\item{ssd1, ssd2}{Numeric vectors, including SSD values.}
}
\value{
A list of three elements, the first two are float, indicating the
RSSD values, and the last one is integer, representing the index of the
dominant prospect.
}
\description{
This function takes two numeric vectors, representing SSD values for the two
prospects, and computes the RSSD values with a comparison showing which one
dominates the other.
}
\details{
The length of `ssd1` and `ssd2` must be equal. Otherwise, an error will be
raised.
}
\examples{
dists = createDistributions(outcome1 = c(1,4,7),
                            outcome2 = c(2,3,5),
                            prob1 = c(1/3,1/3,1/3),
                            prob2 = c(1/6,1/6,2/3))
ssd.obj = ssd.calc(dists)
rssd.test(ssd.obj$ssd1, ssd.obj$ssd2)

}
\seealso{
[ssd(), rssd.calc()] for the parameters.
}
