% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comparisons.R
\name{screen.by.sd}
\alias{screen.by.sd}
\title{Screening by all SD rules}
\usage{
screen.by.sd(data)
}
\arguments{
\item{data}{A data frame, including the results of all SD rules.}
}
\value{
A list, including two elements corresponding to each SD rule.
}
\description{
It checks all available SD rules (i.e. fsd and ssd) based on their results,
and creates the corresponding efficient and inefficient sets.
}
\details{
The input parameter `data` must contain columns corresponding to SD rules.
These columns contain the index of the dominated variable, where 1 or 2 means
the first or the second variable dominates, and 0 means the domination does
not exist. The best practice is to use the output of `compare.paired.distributions`.
}
\examples{
data = compare.paired.distributions(data_ex$gen, rep(1/29,377), data_ex$yield)
sd.sets = screen.by.sd(data)

}
\seealso{
[screen()]
}
