% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Screening.R
\name{screen}
\alias{screen}
\title{A wrapper for computing inefficient and efficient sets}
\usage{
screen(data, test, epsilon.threshold)
}
\arguments{
\item{data}{A data frame, including variable pairs, distributions, and
results of the tests.}

\item{test}{A string that indicates the name of the test.}

\item{epsilon.threshold}{A number that indicates the threshold for considering
the result to be valid or not for ASD tests.}
}
\value{
A list of two elements, efficient and inefficient sets.
}
\description{
A wrapper for computing inefficient and efficient sets
}
\seealso{
[sd.screen(), asd.screen()]
}
