% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SD_ASD_Tests.R
\name{sd.asd.test.all}
\alias{sd.asd.test.all}
\title{Performing SD and ASD tests on distribution pairs}
\usage{
sd.asd.test.all(paired.dists, include.details)
}
\arguments{
\item{paired.dists}{A data frame that includes paired distributions.}

\item{include.details}{A Boolean.}
}
\value{
A data frame that includes the result of tests for each pair.
}
\description{
Performing SD and ASD tests on distribution pairs
}
\details{
`paired.dists` is a data frame including the distributions of all pairs.
`include.details` is a Boolean value. If it is TRUE, the distributions are
included in the output, otherwise they will be discarded.

The output includes the SD and ASD test results, which are the index of the
dominant variable and the epsilon of the ASD tests.
}
\seealso{
[sd.test()]
}
