% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SD_ASD_Tests.R
\name{sd.asd.test}
\alias{sd.asd.test}
\title{Performing all SD and ASD methods on a pair}
\usage{
sd.asd.test(data1, data2)
}
\arguments{
\item{data1, data2}{Data frames corresponding to each element of a pair,
respectively.}
}
\value{
A list of 8 elements indicating the result of each test and the
epsilon values for almost tests.
}
\description{
Performing all SD and ASD methods on a pair
}
\details{
The input parameters are two data frames corresponding to each element of a
pair, respectively. Each data frame has two columns: probability and outcome,
which represents the distribution of the element (variable).

First, a stochastic dominance object is created using the columns of the two
input parameters. Then, all available SD and ASD tests that are implemented
inside the package are performed.

The output list contains 8 elements: `fsd`, `ssd`, `afsd`, `assd.ll`, and
`assd.ths` indicate the index of the dominant element (variable). If the first
variable dominates, they show 1, else they show 2. If neither dominates they
return 0. The other output parameters that end with 'eps' (e.g. `afsd.eps`,
`assd.ll.eps`, and `assd.ths.eps`) show the epsilon corresponding to each
of those ASD tests.
}
\seealso{
[createStochasticDominance(), fsd.test(), ssd.test(), afsd.test(), assd.test()]
}
