% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Screening.R
\name{sd.screen}
\alias{sd.screen}
\title{Create the inefficient set by an SD rule}
\usage{
sd.screen(data, test)
}
\arguments{
\item{data}{A data frame, including variable pairs, distributions, and
results of the tests.}

\item{test}{A string that indicates the name of the test.}
}
\value{
A character vector as the inefficient set of variables based on the
`test` result.
}
\description{
It uses the test result to create the inefficient set.
}
\details{
The inefficient set includes all variable names that are dominated by at
least one other variable. The domination defines as the test result of the
corresponding `test` column.
}
