% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pair_Distributions.R
\name{sorting.variables}
\alias{sorting.variables}
\title{Sorting all variables based on their outcomes}
\usage{
sorting.variables(variable, outcome)
}
\arguments{
\item{variable}{A character vector, includes variable names.}

\item{outcome}{A numeric vector, includes outcome values.}
}
\value{
A character vector, includes all variable names sorted by their
outcomes.
}
\description{
Sorting all variables based on their outcomes
}
