% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSD.R
\name{ssd.calc}
\alias{ssd.calc}
\title{Calculates the SSD values for a pair of prospects.}
\usage{
ssd.calc(dists.obj)
}
\arguments{
\item{dists.obj}{A Distributions object.}
}
\value{
A list, containing two SSD vectors.
}
\description{
Calculates the SSD values for a pair of prospects.
}
\examples{
dists = createDistributions(outcome1 = c(1,4,7),
                            outcome2 = c(2,3,5),
                            prob1 = c(1/3,1/3,1/3),
                            prob2 = c(1/6,1/6,2/3))
ssd.calc(dists)

}
