% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSD.R
\name{ssd}
\alias{ssd}
\title{Calculates the SSD vectors and checks for domination}
\usage{
ssd(outcome, cdf1, cdf2)
}
\arguments{
\item{outcome}{A numeric vector, including all outcomes together in ascending
order.}

\item{cdf1, cdf2}{Numeric vectors, including marginal CDF corresponding to
each prospect.}
}
\value{
A list of two numeric vectors corresponding to SSD values of each
prospect, and an integer value, indicating the dominant prospect.
}
\description{
This function takes three parameters corresponding to the joint distribution
and returns the SSD vectors for each prospect, with their comparison to find
the dominant one.
}
\details{
The length of `outcome`, `cdf1`, and `cdf2` must be equal. If not, an error
is raised.

The output includes three elements: `ssd1`, `ssd2`: two vectors containing
ssd values for the corresponding prospects, and `winner`: an integer that
indicates the dominant prospect (1 or 2), and 0 if the domination fails.
}
\examples{
outcome1 = c(1,4,7)
outcome2 = c(2,3,5)
prob1 = c(1/3,1/3,1/3)
prob2 = c(1/6,1/6,2/3)
obj = fsd(outcome1, outcome2, prob1, prob2)
ssd(obj$outcome, obj$cdf1, obj$cdf2)

}
\seealso{
[fsd()] for the parameters.
}
