% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSD.R
\name{ssd.test}
\alias{ssd.test}
\title{Compares random prospects by SSD}
\usage{
ssd.test(dists.obj)
}
\arguments{
\item{dists.obj}{Distributions object.}
}
\value{
An integer, indicating the index of the dominant prospect.
}
\description{
It compares two random prospects by the second-order stochastic dominance (SSD),
 given the distributions.
}
\details{
If neither prospect dominates the other, it returns 0.
}
\examples{
dists = createDistributions(outcome1 = c(1,4,7),
                            outcome2 = c(2,3,5),
                            prob1 = c(1/3,1/3,1/3),
                            prob2 = c(1/6,1/6,2/3))
 ssd.test(dists)

}
\seealso{
[ssd.calc()] for the parameters.
}
