% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASSD_THS.R
\name{area.below.ssd.calc}
\alias{area.below.ssd.calc}
\title{Calculates area below SSD function}
\usage{
area.below.ssd.calc(outcome, ssd)
}
\arguments{
\item{outcome}{Numeric vector, including outcome values.}

\item{ssd}{Numeric vector, including SSD values.}
}
\value{
Numeric vector, including area below every segment of SSD function.
}
\description{
For every segment, the area below SSD function will be computed.
}
\examples{
outcome = c(1,2,3,4,5,7)
ssd = c(0, 0.33, 0.67, 1, 1.67, 3)
# area.below.ssd.calc(outcome, ssd)

}
\seealso{
[calc.area.below.line()] for area below line.
}
