% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASSD_LL.R
\name{assd.ll.test}
\alias{assd.ll.test}
\title{Compares random prospects by ASSD-LL}
\usage{
assd.ll.test(sd.obj)
}
\arguments{
\item{sd.obj}{StochasticDominance object.}
}
\value{
A list, including the index of the dominant prospect and the epsilon.
}
\description{
It uses the positive and negative areas that are computed by ASSD-LL and the
expected values of the prospects to compare them based on the ASSD-LL rule.
If the violation area ratio is less than 0.5 for a prospect, and its expected
value is larger, it dominates the other by ASSD-LL.
}
\details{
epsilon shows the ratio of the violation. Smaller epsilon means more
decision-makers agree with the result.
}
\examples{
sd = createStochasticDominance(outcome1 = c(1,4,7),
                               outcome2 = c(2,3,5),
                               prob1 = c(1/3,1/3,1/3),
                               prob2 = c(1/6,1/6,2/3))
assd.ll.test(sd)

}
\seealso{
[expected.values(), pos.neg.area.assd.ll(), afsd.test()] for more
details.
}
