% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{expected.values}
\alias{expected.values}
\title{Calculating Expected value}
\usage{
expected.values(outcome1, outcome2, prob1, prob2)
}
\arguments{
\item{outcome1, outcome2}{Numeric vectors, including outcome values.}

\item{prob1, prob2}{Numeric vectors, including probability values.}
}
\value{
A list, including two float elements as the expected value of each
prospect.
}
\description{
It calculates the expected value of both prospects given their
outcomes and probabilities.
}
\details{
All parameters corresponding to the first prospect must have the same length,
meaning `outcome1` and `prob1`. It has to be true for the second prospect as
well. Otherwise, an error will be raised.

The summation of each probability vector, e.g., `prob1` and `prob2` must be
one. Otherwise, an error will be raised.
}
\examples{
 outcome1 = c(1,4,7)
 outcome2 = c(2,3,5)
 prob1 = c(1/3,1/3,1/3)
 prob2 = c(1/6,1/6,2/3)
 expected.values(outcome1, outcome2, prob1, prob2)

}
