% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{expected.values}
\alias{expected.values}
\title{Calculating Expected value}
\usage{
expected.values(dists)
}
\arguments{
\item{dists}{Distributions object.}
}
\value{
A list, including two double elements as the expected value of each
prospect.
}
\description{
It calculates the expected value of both prospects given their distributions.
}
\examples{
dists = createDistributions(outcome1 = c(1,4,7),
                            outcome2 = c(2,3,5),
                            prob1 = c(1/3,1/3,1/3),
                            prob2 = c(1/6,1/6,2/3))
expected.values(dists)

}
