% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{fsd.plot}
\alias{fsd.plot}
\title{Drawing the CDFs}
\usage{
fsd.plot(dists.obj, names = c("1", "2"))
}
\arguments{
\item{dists.obj}{Distributions object.}

\item{names}{A character vector, including the names of prospects in order.}
}
\value{
A list, including plot elements.
}
\description{
It visualizes the CDFs of both prospects.
}
\details{
The parameter `names` only accepts character vector, otherwise an error will
be raised.

The function shows a step plot that includes both CDFs, and return the plot
object for further modifications.
}
\examples{
dists = createDistributions(outcome1 = c(1,4,7),
                            outcome2 = c(2,3,5),
                            prob1 = c(1/3,1/3,1/3),
                            prob2 = c(1/6,1/6,2/3))
 fsd.plot(dists, names = c('First', 'Second'))

}
