% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSD.R
\name{fsd.test}
\alias{fsd.test}
\title{Compares random prospects by FSD}
\usage{
fsd.test(dists.obj)
}
\description{
It compares two random prospects by the first-order stochastic dominance (FSD),
 given the distributions.
}
\details{
@details
 If neither prospect dominates the other, it returns 0.

 @param dists.obj Distributions object.
 @returns An integer, indicating the index of the dominant prospect.
 @examples
dists = createDistributions(outcome1 = c(1,4,7),
                            outcome2 = c(2,3,5),
                            prob1 = c(1/3,1/3,1/3),
                            prob2 = c(1/6,1/6,2/3))
 fsd.test(dists)
}
