% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{modif.outcome.ssd.calc}
\alias{modif.outcome.ssd.calc}
\title{Modify outcome and SSD vectors}
\usage{
modif.outcome.ssd.calc(sd.obj)
}
\arguments{
\item{sd.obj}{StochasticDominance object.}
}
\value{
A list of three elements.
}
\description{
It modifies outcome vector to have all original plus intersection
points in ascending order. The corresponding SSD vectors will also be
returned.
}
\examples{
sd = createStochasticDominance(outcome1 = c(1,4,7),
                               outcome2 = c(2,3,5),
                               prob1 = c(1/3,1/3,1/3),
                               prob2 = c(1/6,1/6,2/3))
modif.outcome.ssd.calc(sd)

}
