% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASSD_LL.R
\name{pos.neg.area.assd.ll}
\alias{pos.neg.area.assd.ll}
\title{Calculates positive and negative area between CDFs for ASSD-LL}
\usage{
pos.neg.area.assd.ll(sd.obj)
}
\arguments{
\item{sd.obj}{StochasticDominance object.}
}
\value{
A list, including two numbers corresponding to the positive and
negative areas, respectively.
}
\description{
It calculates the positive and negative areas between CDFs. The positive is
where both CDF and SSD of the first prospect is larger, and vice versa for
the negative case.
}
\examples{
sd = createStochasticDominance(outcome1 = c(1,4,7),
                               outcome2 = c(2,3,5),
                               prob1 = c(1/3,1/3,1/3),
                               prob2 = c(1/6,1/6,2/3))
pos.neg.area.assd.ll(sd)

}
\seealso{
[modif.outcome.ssd.calc()] for more details.
}
