% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSSD.R
\name{rssd.test}
\alias{rssd.test}
\title{Compares two random prospects by RSSD}
\usage{
rssd.test(sd.obj)
}
\arguments{
\item{sd.obj}{StochasticDominance object.}
}
\value{
A list of three elements, the first two are float, indicating the
RSSD values, and the last one is an integer, representing the index of the
dominant prospect.
}
\description{
It compares two random prospects by RSSD, which is a relaxed version of the
classical SSD. In this method, instead of comparing all SSD values, the
summation of them will be checked, and the prospect having the smaller value
dominates by RSSD.
}
\details{
If no prospect dominates, the returned index for the winner will be 0.
}
\examples{
sd = createStochasticDominance(outcome1 = c(1,4,7),
                               outcome2 = c(2,3,5),
                               prob1 = c(1/3,1/3,1/3),
                               prob2 = c(1/6,1/6,2/3))
rssd.test(sd)

}
