% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{ssd.calc}
\alias{ssd.calc}
\title{Calculates the SSD values for a prospect.}
\usage{
ssd.calc(outcome, cdf)
}
\arguments{
\item{outcome}{Numeric vector, indicating the outcome values.}

\item{cdf}{Numeric vector, indicating the cumulative probabilities.}
}
\value{
Numeric vector, indicating the SSD values.
}
\description{
Calculates the SSD values for a prospect.
}
