% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{ssd.plot}
\alias{ssd.plot}
\title{Drawing the SSD}
\usage{
ssd.plot(outcome, ssd1, ssd2, names = c("1", "2"))
}
\arguments{
\item{outcome}{A numeric vector, including all outcomes in ascending order.}

\item{ssd1, ssd2}{Numeric vectors, including SSD values corresponding to each
prospects.}
}
\value{
A list, including plot elements.
}
\description{
It takes three parameters to visualize the SSD values of both prospects.
}
\details{
The length of all parameters must be equal. Otherwise, an error will be
raised.

The function shows the plot and return a plot object for further modification.
}
\examples{
outcome1 = c(1,4,7)
outcome2 = c(2,3,5)
prob1 = c(1/3,1/3,1/3)
prob2 = c(1/6,1/6,2/3)
obj.fsd = fsd(outcome1, outcome2, prob1, prob2)
obj.ssd = ssd(obj.fsd$outcome, obj.fsd$cdf1, obj.fsd$cdf2)
plot_ssd(obj.fsd$outcome, obj.ssd$ssd1, obj.ssd$ssd2)

}
\seealso{
[ssd()] for the parameters.
}
